feat(deps): bump reth to glamsterdam-devnet-7 and integrate new revm
Bump reth git deps to 4f36452 (glamsterdam-devnet-7) and mirror its
patch section: revm 0c688c68, revm-inspectors 1f82bd2 (0.41.2),
alloy-evm 1e54097, reth-core 8cc57a8.

Adapt to the revm EIP-2780 runtime gas phase restructuring while
keeping EIP-2780 disabled on Tempo:
- thread the transaction-level Gas through the AA single/multi-call
  execution paths and commit the runtime phase checkpoint upfront
- move the T0+ intrinsic-gas-limit OOG check into
  run_without_catch_error overrides and rebuild the tx-level gas after
  pre-execution so the 2D-nonce CREATE surcharge still applies
- charge the EIP-8037 state-gas intrinsics (per-auth and CREATE) in
  validate_initial_tx_gas now that upstream defers them to the runtime
  phase
- follow the tx_eip7702_regular_gas/tx_eip7702_regular_refund GasId
  renames and the signed PrecompileOutput::state_gas_used

Adapt to reth API changes: BuildArguments::state_root_handle rename
with the take-once state hook and Arc'd hashed state, the new
EthSubscriptions helper trait, and the GetSnap eth request variant.

+++ crates/precompiles/src/dispatch.rs
@@ -106,7 +106,7 @@ fn fill_state_gas(output: &mut PrecompileOutput, storage: &StorageCtx) {
         if output.is_success() {
             // On success: parent takes the child's final reservoir.
             output.reservoir = storage.reservoir();
-            output.state_gas_used = storage.state_gas_used();
+            output.state_gas_used = storage.state_gas_used() as i64;
         } else {
             // On revert or halt: state changes are undone, so ALL state gas returns
             // to the parent's reservoir.
